fix inj_ticks status, wasn't being updated and now uint16_t

+++ src/ecu.h
@@ -32,7 +32,7 @@ typedef struct _ecu_status
   uint16_t engine_start_ms;
   uint16_t engine_prime_ms;
   // outputs
-  uint8_t inj_ticks;
+  uint16_t inj_ticks;
   uint16_t thr_in;
   uint16_t pwm0_out;
   uint16_t pwm1_out;
